Add Tabs to detail page
