+++ src/detail-page/components/JobsTab.tsx
@@ -4,11 +4,9 @@ interface JobProps {
   content: any;
 }
 
-export const JobsTab: React.FC<JobProps> = ({ content }) => {
-  return (
-    <div>
-      {/* Should be replace with actual content */}
-      {content}
-    </div>
-  );
-};
+export const JobsTab: React.FC<JobProps> = ({ content }) => (
+  <div>
+    {/* Should be replace with actual content */}
+    {content}
+  </div>
+);
