+++ src/detail-page/components/LearnTab.tsx
@@ -4,11 +4,9 @@ interface LearnProps {
   content: any;
 }
 
-export const LearnTab: React.FC<LearnProps> = ({ content }) => {
-  return (
-    <div>
-      {/* Should be replace with actual content */}
-      {content}
-    </div>
-  );
-};
+export const LearnTab: React.FC<LearnProps> = ({ content }) => (
+  <div>
+    {/* Should be replace with actual content */}
+    {content}
+  </div>
+);
