+++ src/detail-page/components/ResourcesTab.tsx
@@ -4,11 +4,9 @@ interface ResourcesProps {
   content: any;
 }
 
-export const ResourcesTab: React.FC<ResourcesProps> = ({ content }) => {
-  return (
-    <div>
-      {/* Should be replace with actual content */}
-      {content}
-    </div>
-  );
-};
+export const ResourcesTab: React.FC<ResourcesProps> = ({ content }) => (
+  <div>
+    {/* Should be replace with actual content */}
+    {content}
+  </div>
+);
