+++ src/layout/theme.ts
@@ -54,9 +54,6 @@ export const theme = createMuiTheme({
     primary: {
       main: '#725ec0',
     },
-    secondary: {
-      main: green[500],
-    },
     text: {
       primary: '#222222',
       disabled: '#22222233', // Text color with opacity 30%
